refactor: moved `MixinValue` type to feature
