+++ packages/core/src/features/index.ts
@@ -12,7 +12,7 @@ export * as STVar from './st-var';
 export type { VarSymbol } from './st-var';
 
 export * as STMixin from './st-mixin';
-export type { RefedMixin } from './st-mixin';
+export type { RefedMixin, MixinValue } from './st-mixin';
 
 export * as CSSClass from './css-class';
 export type { ClassSymbol } from './css-class';
