+++ packages/core/src/index.ts
@@ -14,6 +14,7 @@ export type {
     Imported,
     KeyframesSymbol,
     RefedMixin,
+    MixinValue,
     StylableDirectives,
     StylableSymbol,
     VarSymbol,
@@ -79,7 +80,6 @@ export {
     ArgValue,
     ExtendsValue,
     MappedStates,
-    MixinValue,
     ReportWarning,
     SBTypesParsers,
     STYLABLE_NAMED_MATCHER,
